+++ 後端/HomeBackProject/HomeBackProject/Models/PeopleRankData.cs
@@ -15,6 +15,7 @@ namespace HomeBackProject.Models
     using System.ComponentModel.DataAnnotations;
     using MetaDataHomeBackProject.Models;
     [MetadataType(typeof(MetaPeopleRankData))]
+
     public partial class PeopleRankData
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
@@ -23,7 +24,7 @@ namespace HomeBackProject.Models
             this.PeopleData = new HashSet<PeopleData>();
         }
     
-        public byte HomeTSaleStateID { get; set; }
+        public int HomeTSaleStateID { get; set; }
         public string PeopleRank { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
