+++ 後端/HomeBackProject/HomeBackProject/Models/SaleTypeData.cs
@@ -26,7 +26,7 @@ namespace HomeBackProject.Models
             this.TerritoryData = new HashSet<TerritoryData>();
         }
     
-        public byte SaleStateID { get; set; }
+        public int SaleStateID { get; set; }
         public string SaleState { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
