+++ 後端/HomeBackProject/HomeBackProject/Models/TerritoryTypeData.cs
@@ -25,7 +25,7 @@ namespace HomeBackProject.Models
             this.TerritoryData = new HashSet<TerritoryData>();
         }
     
-        public byte TerritorySerialID { get; set; }
+        public int TerritorySerialID { get; set; }
         public string TerritoryTypeSelect { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
