+++ 後端/HomeBackProject/HomeBackProject/Models/TerritoryadjacentTypeData.cs
@@ -15,9 +15,10 @@ namespace HomeBackProject.Models
     using System.ComponentModel.DataAnnotations;
     using MetaDataHomeBackProject.Models;
     [MetadataType(typeof(MetaTerritoryadjacentTypeData))]
+
     public partial class TerritoryadjacentTypeData
     {
-        public byte TerritoryAdjacentStateID { get; set; }
+        public int TerritoryAdjacentStateID { get; set; }
         public string TerritoryAdjacentName { get; set; }
         public string TerritoryID { get; set; }
